pyproject.toml
[tool.poetry]
name = "sig-health-slackbot"
version = "0.0.1"
description = "A slackbot which make our teammates healthier"
authors = ["thehoodieguy"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.1"
djangorestframework = "^3.11.0"
daphne = "^2.4.1"
psycopg2-binary = "^2.8.4"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pre-commit = "^1.20.0"
black = "^19.10b0"
isort = "^4.3.21"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
